removed redundant test files
